+++ packages/stream_video/lib/src/webrtc/rtc_audio_api/rtc_audio_stub.dart
@@ -25,6 +25,8 @@ Future<void> resumeAudioPlayback() {
 
 bool get isAudioPlaybackBlocked => throw UnimplementedError();
 
+bool get isAudioPlaybackStalled => throw UnimplementedError();
+
 bool checkIfAudioOutputChangeSupported() {
   throw UnimplementedError();
 }